% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_path}
\alias{add_path}
\title{Overlays path onto existing plot}
\usage{
add_path(plot, df, path, slider = 0)
}
\arguments{
\item{plot}{A \code{ggplot} object. Plot the path is to be overlaid upon.}

\item{df}{A data frame containing columns named \code{x} and \code{y} containing the
coordinates of the points in the plot.}

\item{slider}{A non-negative numeric denoting which segment in the path to
highlight. If \code{slider = 0}, no segment will be highlighted.}

\item{A}{named list returned by \code{get_shortest_path()}.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{add_path()} returns a plot with the path, specified by \code{df} and \code{path},
overlaid. \code{slider} can be used to highlight a specific segment of the plot.
}
