% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_test.R, R/mst_test_brush.R
\name{count_crossings}
\alias{count_crossings}
\alias{count_crossings_brush}
\title{Count the number of crossings between two clusters in the inputted tree}
\usage{
count_crossings(mst, path, cluster)

count_crossings_brush(mst, g1, g2)
}
\arguments{
\item{mst}{An \code{igraph} object.}

\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}

\item{cluster}{A vector of length \code{vcount(mst)} with cluster labels.}

\item{g1, g2}{A numerical vector of indices of the points in each group.}
}
\value{
A numeric.
}
\description{
Given two clusters and a tree, a simplified subtree is created (see
\code{\link[=simplify_sub_mst]{simplify_sub_mst()}}). The number of crossings between the two clusters in
the simplified subtree is then counted.
}
