% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_medoid_mst.R
\name{get_medoid_mst}
\alias{get_medoid_mst}
\title{Constructs the minimal subtree containing all cluster medoids}
\usage{
get_medoid_mst(Z_dist, mst, cluster)
}
\arguments{
\item{Z_dist}{A \code{dist} object or numerical distance matrix.}

\item{mst}{An \code{igraph} object.}

\item{cluster}{A vector of length \code{nrows(as.matrix(Z_dist))}.}
}
\value{
An \code{igraph} object. The returned tree has vertex attribute \code{medoid}
denoting which cluster each point is the medoid of. For non-medoid points,
this attribute is NA.
}
\description{
Uses \code{\link[=get_medoids]{get_medoids()}} to calculate cluster medoids, then uses \code{\link[=get_subtree]{get_subtree()}}
to get the minimal subtree.
}
\examples{
Z <- matrix(rnorm(20), nrow=5, ncol=4)
Z_dist <- dist(Z)
mst = get_mst(Z_dist)
cluster <- c(1,1,2,2,2)
get_medoid_mst(Z_dist, mst, cluster)
}
