% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_medoid_mst.R
\name{get_medoids}
\alias{get_medoids}
\title{Get the medoid of each cluster}
\usage{
get_medoids(Z_dist, cluster)
}
\arguments{
\item{Z_dist}{A \code{dist} object or numerical distance matrix.}

\item{cluster}{A vector of length \code{nrow(as.matrix(dist))}.}
}
\value{
A numerical vector of medoid indices.
}
\description{
Returns the indices of the medoid of each cluster in the data set.
}
