% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mst.R
\name{get_mst}
\alias{get_mst}
\title{Compute the minimum spanning tree of set of points}
\usage{
get_mst(Z_dist)
}
\arguments{
\item{Z_dist}{A \code{dist} object or distance matrix.}
}
\value{
An \code{igraph} object.
}
\description{
Returns an \code{igraph} object containing the minimum spanning tree. Constructs a
graph using \code{\link[igraph:graph_from_adjacency_matrix]{igraph::graph_from_adjacency_matrix()}}, then calculates the minimum
spanning tree using \code{\link[igraph:mst]{igraph::mst()}}.
}
