% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_test.R
\name{get_path_endpts}
\alias{get_path_endpts}
\title{Retrieves the id's of the path endpoints.}
\usage{
get_path_endpts(path, cluster, id)
}
\arguments{
\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}

\item{cluster}{A vector with cluster labels.}

\item{id}{A vector of point id's.}
}
\value{
A list with the following components:
\item{from}{The id of the first point in the path.}

\item{to}{The id of the last point in the path.}

\item{same}{A Boolean denoting whether the first and last points of the path
belong to the same cluster.}
}
\description{
This function retrieves the id's of the path endpoints, which may be
different from their indices.
}
