% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_weights.R
\name{get_path_weights}
\alias{get_path_weights}
\title{Get path weights}
\usage{
get_path_weights(path)
}
\arguments{
\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}
}
\value{
A numerical vector of length \code{length(path$epath)}.
}
\description{
Returns a vector of weights for a given path.
}
