% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection.R, R/get_projection_brush.R
\name{get_projection}
\alias{get_projection}
\alias{get_projection_brush}
\title{Calculate the projection of the selected points using the PCA-rCCA method}
\usage{
get_projection(Z, path, cluster, dim, degree)

get_projection_brush(Z, path, g1, g2, cluster, dim, degree)
}
\arguments{
\item{Z}{A numerical matrix containing the high-dimensional data.}

\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}

\item{cluster}{A vector of length \code{nrow(Z)} with cluster labels.}

\item{dim}{A numeric greater than or equal to 2.}

\item{degree}{A positive numeric.}

\item{g1, g2}{A numerical vector of the indices of the points in each group.}
}
\value{
A list with the following components:
\item{projected_pts}{A numerical matrix with two columns containing the
projected coordinates.}

\item{ids}{A numerical vector containing the indices of the selected points.}

\item{path_ids}{A numerical vector containing the indices of the path points.}

\item{var_explained}{A numeric. The proportion of variance retained during
the PCA projection step.}
}
\description{
The selected points are first projected down to \code{dim} dimensions using PCA.
They are then projected down to two dimensions using rCCA. The comparison
matrix is the design matrix for a polynomial of degree \code{degree}. The points
are selected in one of two ways:
\itemize{
\item \code{get_projection()} selects the path points along with all points belonging
to the same cluster as one of path's endpoints.
\item \code{get_projection_brush()} selects the path points along with the points
belonging to the two groups selected by the user.
}
}
