% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection.R
\name{get_projection_inner}
\alias{get_projection_inner}
\title{Internal function for \code{\link[=get_projection]{get_projection()}} and \code{\link[=get_projection_brush]{get_projection_brush()}}}
\usage{
get_projection_inner(Z, path, dim, degree, path_ids, path_pts, ids)
}
\arguments{
\item{Z}{A numerical matrix containing the high-dimensional data.}

\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}

\item{dim}{A numeric greater than or equal to 2.}

\item{degree}{A positive numeric.}

\item{path_ids}{A numerical vector containing the indices of the points along
the path.}

\item{path_pts}{A numeric matrix containing the coordinates of the points
along the path.}

\item{ids}{A numerical vector containing the indices of the points to
project.}
}
\value{
A list with the following components:
\item{projected_pts}{A numerical matrix with two columns containing the
projected coordinates.}

\item{ids}{A numerical vector containing the indices of the selected points.}

\item{path_ids}{A numerical vector containing the indices of the path points.}

\item{var_explained}{A numeric. The proportion of variance retained during
the PCA projection step.}
}
\description{
Internal function used by \code{\link[=get_projection]{get_projection()}} and \code{\link[=get_projection_brush]{get_projection_brush()}} to
calculate the PCA-rCCA projection.
}
