% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shortest_path.R
\name{get_shortest_path}
\alias{get_shortest_path}
\title{Computes shortest path between two points in a graph}
\usage{
get_shortest_path(graph, from, to)
}
\arguments{
\item{graph}{An \code{igraph} object.}

\item{from}{A numeric.}

\item{to}{A numeric.}
}
\value{
A named list with the \code{epath} and \code{vpath} returned by
\code{\link[igraph:distances]{igraph::shortest_paths()}}. See \code{\link[igraph:distances]{igraph::distance_table()}} for reference.
}
\description{
Wrapper for \code{\link[igraph:distances]{igraph::shortest_paths()}}. Returns the shortest path in \code{graph}
between vertices \code{from} and \code{to}.
}
\details{
The function assumes the inputted graph is a tree.
}
