% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_test.R
\name{get_sim_param}
\alias{get_sim_param}
\title{Calculate the parameters for crossings simulation}
\usage{
get_sim_param(Z1, Z2, keep)
}
\arguments{
\item{Z1}{A numerical matrix containing the data for group 1.}

\item{Z2}{A numerical matrix containing the data for group 2.}

\item{keep}{A numeric between 0 and 1. The proportion of variance to retain
when truncating dimensions.}
}
\value{
A list with the following components:
\item{num_pts}{A numeric. The number of points to be sampled.}

\item{side_lengths}{A numerical vector. The side lengths of the hyperrectangle
to be sampled from.}
}
\description{
The density of a cluster is estimated to be the volume of a hyperrectangle
with side lengths equal to sqrt(12) times the singular values of the
cluster. This is because a uniform distribution on the hyperrectangle will
have variances equal to the variances of the cluster in the directions of the
principal components. The simulation parameters are chosen to match the
group with the lower density.
}
