% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_medoid_mst.R
\name{get_subtree}
\alias{get_subtree}
\title{Get the minimum subtree containing the specified vertices}
\usage{
get_subtree(tree, points)
}
\arguments{
\item{tree}{An \code{igraph} object}

\item{points}{A numeric vector of vertex indices.}
}
\value{
An \code{igraph} object.
}
\description{
The subtree contains all of the specified vertices along with the shortest
paths between each pair of them. Uses \code{\link[igraph:subgraph]{igraph::induced_subgraph()}}.
}
