% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_plot.R, R/meta_data_plot_brush.R
\name{meta_data_plot}
\alias{meta_data_plot}
\alias{meta_data_plot_brush}
\title{Plot meta data}
\usage{
meta_data_plot(Z, path, cluster, meta_data, feature)

meta_data_plot_brush(Z, g1, g2, meta_data, feature)
}
\arguments{
\item{Z}{A numerical matrix containing the high-dimensional data.}

\item{path}{A named list returned by \code{get_shortest_path()}.}

\item{cluster}{A vector of length \code{nrow(Z)} with cluster labels.}

\item{meta_data}{A data frame with a number of rows equal to \code{nrow(Z)}. For
presenting extra meta data.}

\item{feature}{A string containing the name of the feature to be displayed.}

\item{g1, g2}{A numerical vector of the indices of the points each group.}
}
\value{
A \code{ggplot} or \code{gtable} object.
}
\description{
Plots the meta data for the selected groups. Categorical data is plotted
with pie charts, while numerical data is plotted with box plots.
}
