% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNIST_id}
\alias{MNIST_id}
\title{MNIST ids}
\format{
An object of class \code{integer} of length 2000.
}
\usage{
MNIST_id
}
\description{
A vector of ids used to keep track of which MNIST images were randomly
sampled.
}
\keyword{datasets}
