% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNIST_kmeans_cluster}
\alias{MNIST_kmeans_cluster}
\title{Output of k-means clustering on MNIST data}
\format{
An object of class \code{integer} of length 2000.
}
\usage{
MNIST_kmeans_cluster
}
\description{
A vector of class labels computed using \code{\link[=kmeans]{kmeans()}} on \link{MNIST_pca} with 10 clusters, one for
each digit.
}
\keyword{datasets}
