% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNIST_labels}
\alias{MNIST_labels}
\title{MNIST class labels}
\format{
An object of class \code{integer} of length 2000.
}
\usage{
MNIST_labels
}
\description{
A vector of MNIST class labels pulled from the \code{dslabs} package for \link{MNIST_pca}. The labels
correspond to the handwritten digits.
}
\keyword{datasets}
