% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNIST_low}
\alias{MNIST_low}
\title{t-SNE embedding of MNIST data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2000 rows and 2 columns.
}
\usage{
MNIST_low
}
\description{
A 2-dimensional t-SNE embedding of \link{MNIST_pca} computed using \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}.
}
\keyword{datasets}
