% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNIST_pca}
\alias{MNIST_pca}
\title{PCA-processed MNIST data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2000 rows and 300 columns.
}
\usage{
MNIST_pca
}
\description{
MNIST data set pulled from the \code{dslabs} package. \code{\link[=prcomp]{prcomp()}} used to compute a
300-dimensional PCA embedding, then 2,000 rows were randomly sampled.
}
\keyword{datasets}
