% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_test.R, R/mst_test_brush.R
\name{mst_test}
\alias{mst_test}
\alias{mst_test_brush}
\title{Runs the MST test for two clusters}
\usage{
mst_test(sim_crossings, num_crossings, endpts)

mst_test_brush(sim_crossings, num_crossings)
}
\arguments{
\item{sim_crossings}{A numerical vector of simulated numbers of crossings.}

\item{num_crossings}{The actual number of crossings.}

\item{endpts}{A named list returned by \code{\link[=get_path_endpts]{get_path_endpts()}}.}
}
\description{
Calculates a p-value from the simulated numbers of crossings and the actual
number of crossings, then prints the results.
}
