% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_projection.R,
%   R/plot_2d_projection_brush.R
\name{plot_2d_projection}
\alias{plot_2d_projection}
\alias{plot_2d_projection_brush}
\title{Constructs plot of projected points}
\usage{
plot_2d_projection(
  mst,
  cluster,
  id,
  projected_pts,
  ids,
  path_ids,
  var_explained,
  degree,
  slider,
  adjust,
  show_all_edges
)

plot_2d_projection_brush(
  mst,
  cluster,
  id,
  g1,
  g2,
  projected_pts,
  ids,
  path_ids,
  var_explained,
  degree,
  slider,
  adjust,
  show_all_edges,
  color_choice
)
}
\arguments{
\item{mst}{An \code{igraph} object.}

\item{cluster}{A vector of length \code{nrow(projected_pts)} containing the
cluster labels.}

\item{id}{A numerical vector containing the labels of the points.}

\item{projected_pts}{A numerical matrix with two columns.}

\item{ids}{A numerical vector containing the indices of the projected points.}

\item{path_ids}{A numerical vector containing the indices of the path points.}

\item{var_explained}{A numeric.}

\item{degree}{A numeric. The degree of the polynomial design matrix during
rCCA when computing the projection.}

\item{slider}{A numeric. Determines which segment of the path should be
highlighted in the plot. If \code{slider = 0}, no segment will be highlighted.}

\item{adjust}{A numeric. The bandwidth of the overlaid density estimation. If
\code{adjust = 0}, the estimated density will not be overlaid the plot.}

\item{show_all_edges}{"Show" or "Hide". If "Show", all edges of the minimum
spanning tree will be drawn.}

\item{g1, g2}{A numerical vector containing the indices of the points in each
group.}

\item{color_choice}{"Original Coloring" or "Group Coloring". Determines if
the colors of the projected points are returned or if they're colored
according to the user-selected groups.}
}
\value{
A named list containing the plot \code{p} and the variance explained
\code{var_explained}.
}
\description{
Uses the output of \code{\link[=get_projection]{get_projection()}} or \code{\link[=get_projection_brush]{get_projection_brush()}} to
construct a plot of the projected points. Overlaid on the plot are the
projected path and estimated density (optional). In addition, the use may
choose to display all edges of the minimum spanning tree, not just the path.
}
\details{
The variance explained is not written on the plot because it would get
removed by \code{\link[plotly:ggplotly]{plotly::ggplotly()}}. Instead the variance explained is passed along, then
added after the plot is converted into a \code{plotly} object.
}
