% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R, R/plot_heatmap_brush.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap_brush}
\title{Construct the heatmap of high-dimensional points}
\usage{
plot_heatmap(Z, path, cluster, col_names)

plot_heatmap_brush(Z, g1, g2, col_names)
}
\arguments{
\item{Z}{A numerical matrix containing the high-dimensional data.}

\item{path}{A named list returned by \code{get_shortest_path()}.}

\item{cluster}{A vector of length \code{nrow(Z)} with cluster labels.}

\item{col_names}{A vector of length \code{col(Z)} with column names.}

\item{g1, g2}{A numerical vector of the indices of the points each group.}
}
\value{
A \code{ComplexHeatmap} object.
}
\description{
Creates a heatmap of the high-dimensional points. The rows are split
according to
\itemize{
\item class for \code{\link[=plot_heatmap]{plot_heatmap()}}, and
\item user-defined groups for \code{\link[=plot_heatmap_brush]{plot_heatmap_brush()}}.
}
}
