% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_medoid_mst.R
\name{plot_medoid_mst}
\alias{plot_medoid_mst}
\title{Overlays the minimal subtree containing all cluster medoids}
\usage{
plot_medoid_mst(plot, df, Z_dist, tree)
}
\arguments{
\item{plot}{A \code{ggplot} object.}

\item{df}{A data frame containing columns named \code{x} and \code{y} containing the
coordinates of the points in the plot.}

\item{Z_dist}{A \code{dist} object or numerical distance matrix.}

\item{tree}{An \code{igraph} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Uses \code{\link[=get_medoid_mst]{get_medoid_mst()}} to construct the minimal subtree containing the
cluster medoids, then overlays it over \code{plot}.
}
