% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_test.R, R/mst_test_brush.R
\name{sim_crossings}
\alias{sim_crossings}
\alias{sim_crossings_brush}
\title{Simulate the null distribution for number of crossings}
\usage{
sim_crossings(Z, path, cluster, b, keep = 0.9, parallel = FALSE)

sim_crossings_brush(Z, g1, g2, cluster, b, keep = 0.9, parallel = FALSE)
}
\arguments{
\item{Z}{A numerical matrix containing the high-dimensional data.}

\item{path}{A named list returned by \code{\link[=get_shortest_path]{get_shortest_path()}}.}

\item{cluster}{A vector of length \code{nrow(Z)} with cluster labels.}

\item{b}{A positive numeric. The number of simulations to run.}

\item{keep}{A numeric between 0 and 1. The proportion of variance to retain
when truncating dimensions.}

\item{parallel}{A Boolean denoting whether parallel computing should be
used. Parallel computing is implemented via \code{\link[parallel:mclapply]{parallel::mclapply()}}, which is
not available on Windows.}

\item{g1, g2}{A numerical vector of indices of the points in each group.}
}
\value{
A numerical vector.
}
\description{
The MST is calculated on uniform sample drawn from a hyperrectangle. The
number of edges crossing a bisecting hyperplane is recorded, and this
process is repeated multiple times. The dimensions of the hyperrectangle are
chosen to match the data.
}
