% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_sub_mst.R
\name{simplify_sub_mst}
\alias{simplify_sub_mst}
\title{Simplify subtree for count_crossings()}
\usage{
simplify_sub_mst(mst, g1, g2, cluster)
}
\arguments{
\item{mst}{An \code{igraph} object.}

\item{g1, g2}{A numerical vector of indices of the points in each group.}

\item{cluster}{A vector of length \code{vcount(mst)} with cluster labels.}
}
\value{
An \code{igraph} object.
}
\description{
Computes the minimal subtree containg g1 and g2, then simplifies it in
preparation for \code{\link[=count_crossings]{count_crossings()}}. First, paths of non-groups points are
condensed into one edge. Then, edges conned non-group points are collapsed.
}
