% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_cell_labels}
\alias{Wong_cell_labels}
\title{Wong cell type labels}
\format{
An object of class \code{character} of length 3000.
}
\usage{
Wong_cell_labels
}
\description{
A vector of cell type labels for \link{Wong_high}.
}
\keyword{datasets}
