% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_high}
\alias{Wong_high}
\title{Wong et al. (2016) data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3000 rows and 40 columns.
}
\usage{
Wong_high
}
\description{
Mass cytometry data from \href{https://doi.org/10.1016/j.immuni.2016.07.007}{Wong et al. (2016)}.
The data was processed according to \href{https://www.nature.com/articles/doi:10.1038/nbt.4314.epdf?sharing_token=G7k__zflxFlYhRLIRySpAtRgN0jAjWel9jnR3ZoTv0NFmV-pW0O8ddmuAfWKT7uTCry4tCzWyL3Wzdfnz5B_0wyyWhwrSsy-9fWd7f0j5VlMUrbzKpcqFWIoL27QL8fWgCGw_KLJuImgYWttt7L_PQ\%3D\%3D}{Becht et al. (2018)},
then 3,000 samples were randomly selected.
}
\keyword{datasets}
