% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_kmeans}
\alias{Wong_kmeans}
\title{Output of k-means clustering on Wong et al. (2016) data}
\format{
An object of class \code{integer} of length 3000.
}
\usage{
Wong_kmeans
}
\description{
A vector of class labels computed using \code{\link[=kmeans]{kmeans()}} with 10 clusters on \link{Wong_high}.
}
\keyword{datasets}
