% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_low}
\alias{Wong_low}
\title{UMAP embedding of Wong et al. (2016)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3000 rows and 2 columns.
}
\usage{
Wong_low
}
\description{
A 2-dimensional UMAP embedding of \link{Wong_high} computed using \code{\link[umap:umap]{umap::umap()}}.
}
\keyword{datasets}
