% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_organ_labels}
\alias{Wong_organ_labels}
\title{Wong tissue type labels}
\format{
An object of class \code{character} of length 3000.
}
\usage{
Wong_organ_labels
}
\description{
A vector of tissue type labels for \link{Wong_high}.
}
\keyword{datasets}
