% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_medoid.R
\name{get_medoid}
\alias{get_medoid}
\title{Get medoid of a cluster}
\usage{
get_medoid(X_dist, g)
}
\arguments{
\item{X_dist}{A \code{dist} object or distance matrix.}

\item{g}{A numerical vector of indices that denotes which group of points to get the
medoid of.}
}
\value{
A numeric.
}
\description{
\code{get_medoid()} returns the index of the medoid of the cluster of points
given by \code{g}.
}
\examples{
X <- matrix(rnorm(40), nrow=10, ncol=4)
g <- c(1,2,3)
get_medoid(dist(X), g)
}
