% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the dimension reduction tool}
\usage{
run_app(Z, X, cluster, id = NULL)
}
\arguments{
\item{Z}{A numerical matrix. The high-dimensional data.}

\item{X}{A numerical matrix satisfying \code{nrow(X) = nrow(Z)} and
\code{ncol(X) = 2}.}

\item{cluster}{A numerical vector or factor of length \code{nrow(Z)}.}

\item{id}{A vector of length \code{nrow(Z)}. If \code{id = NULL}, the ids will be set
to the indices of the points.}
}
\description{
Launches the dimension reduction tool. The required inputs are the
high-dimensional data matrix, the low-dimensional data matrix, and the clustering.
Optionally, a vector of ids can be given to annotate the points.
}
\examples{
library(MASS)
library(Rtsne)

# construct high-dimensional data, three clusters
Z1 <- mvrnorm(n=30, mu=c(0,0,0,0,0), Sigma=diag(rep(1,5)))
Z2 <- mvrnorm(n=30, mu=c(0,2,0,0,0), Sigma=diag(rep(1,5)))
Z3 <- mvrnorm(n=30, mu=c(0,0,2,0,0), Sigma=diag(rep(1,5)))
Z <- rbind(Z1, Z2, Z3)

# compute low-dimensional embedding using t-SNE
X <- Rtsne(Z, dims=2, perplexity=30)$Y

# compute k-means clustering
cluster <- kmeans(Z, centers=3, nstart=10)$cluster

#launch tool
run_app(Z, X, cluster)
}
