% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Launch dimension reduction tool examples}
\usage{
run_example(
  example = c("MNIST", "Wong"),
  cluster = c("real", "kmeans"),
  parallel = FALSE
)
}
\arguments{
\item{example}{A character string indicating which example data set to use.
See \link{MNIST_pca} and \link{Wong_high}.}

\item{cluster}{A character string indicating the clustering to project onto
the low-dimensional embedding. Possible values of cluster are "real" and
"kmeans".}

\item{parallel}{A Boolean indicating whether parallel computing should be
used. The implementation uses \code{\link[parallel:mclapply]{parallel::mclapply()}}, which is not available
on Windows.}
}
\value{
None
}
\description{
The current implementation contains one example on the MNIST data set. For
each example, the tool can be launched visualizing the true class labels or
a computed k-means clustering.
}
\examples{
# parallel computing is not available on Windows
if (interactive()) {
  run_example(example="MNIST", cluster="real", parallel=FALSE)
}

}
