% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wong_high}
\alias{Wong_high}
\title{Wong et al. (2016) data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3000 rows and 40 columns.
}
\usage{
Wong_high
}
\description{
Mass cytometry data from \href{https://doi.org/10.1016/j.immuni.2016.07.007}{Wong et al. (2016)}.
The data was processed according to \href{https://rdcu.be/edjyF}{Becht et al. (2018)},
then 3,000 samples were randomly selected.
}
\keyword{datasets}
